NoteOn() reset for data sources -BindManager.cs:  set the data source and member name (property name) -CachedEnvelope.cs:  Move reset functionality to its own method -CachedEnvelopePoint.cs:  Change dynamic switch expression to a traditional switch statement -Interfaces.cs:  IBindableDataConsumer has default NoteOn implementation -TrackerEnvelope.cs:  fixed invalid cast -Operator.cs:  Implement IBindableDataConsumer default interface for NoteOn() -Test2.cs:  switch expressions to access the correct datatype of the items inside the BoundEnvelopes container -BasePanel.gd:  Fixed erroneous print message

+++ PhaseEngine/BindAutomation/TrackerEnvelope.cs
@@ -254,8 +254,8 @@ namespace PhaseEngine
                     o.AddPrim("maxValue", (float)Convert.ChangeType(maxValue, typeof(float)));
                     break;
                 default:
-                    o.AddPrim("minValue", (float)Convert.ChangeType(minValue, typeof(int)));
-                    o.AddPrim("maxValue", (float)Convert.ChangeType(maxValue, typeof(int)));
+                    o.AddPrim("minValue", (int)Convert.ChangeType(minValue, typeof(int)));
+                    o.AddPrim("maxValue", (int)Convert.ChangeType(maxValue, typeof(int)));
                     break;
             }
